refactor(sessions): move canonical session resolution
