+++ scripts/check-session-accessor-boundary.mjs
@@ -137,7 +137,7 @@ export const migratedSessionAccessorFiles = new Set([
   "src/gateway/boot.ts",
   "src/gateway/server-methods/artifacts.ts",
   "src/gateway/server-methods/chat.ts",
-  "src/gateway/sessions-resolve.ts",
+  "src/sessions/session-resolve.ts",
   "src/gateway/server-methods/sessions-files.ts",
   ...gatewaySessionServerMethodFiles,
   "src/gateway/server-session-events.ts",
